+++ src/main/java/com/example/nutriplan_backend/repository/GrupoAlimentoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GrupoAlimentoRepository extends JpaRepository<GrupoAlimento, Long> {
+public interface GrupoAlimentoRepository extends JpaRepository<GrupoAlimento, Integer> {
 }
