+++ src/main/java/com/example/nutriplan_backend/repository/ObjetivoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ObjetivoRepository extends JpaRepository<Objetivo, Long>{
+public interface ObjetivoRepository extends JpaRepository<Objetivo, Integer>{
 
 }
